requirements.txt
flask~=1.1.2
flask-restx==0.2.0
distro==1.5.0
numpy==1.19.5
pandas==1.2.0
pickle5==0.0.11
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
tabula-py==2.2.0
tqdm==4.56.0
